petget: fix scripts to work easyvoid

+++ easyos/easy-code/rootfs-skeleton/usr/local/petget/pkg_chooser.sh
@@ -47,6 +47,7 @@
 #20230711 when called from /usr/bin/*.install script, has passed param "gen-tmp-files-only"
 #20230914 void: update pkg db every time run pkgget.
 #20230914 stupid grep: "grep: warning: stray \ before -" use busybox grep. no, grep -P works. no, in case only have busybox grep, it doesn't understand -P
+#20240228 when easyvoid has pkgget frontend for xbps, no need to update pkg db at startup.
 
 #/usr/local/petget/service_pack.sh & #121125 offer download Service Pack.
 
@@ -73,6 +74,13 @@ mkdir -p /var/local/petget
 #this must be after running apt-setup
 . /root/.packages/PKGS_MANAGEMENT #has PKG_REPOS_ENABLED, PKG_NAME_ALIASES
 
+case "$DISTRO_TARGETARCH" in #20240228
+ amd64) xARCH='x86_64' ;;
+ *)     xARCH="$DISTRO_TARGETARCH" ;;
+esac
+#actually, do this in installpreview.sh, but probably good do it up-front...
+export XBPS_ARCH="$xARCH"
+
 #190813 aborted update of db's resulted in *pre files getting left behind...
 for aPRE in `find /root/.packages -mindepth 1 -maxdepth 1 -type f -name 'Packages-*pre'`
 do
@@ -80,19 +88,22 @@ do
 done
 
 if [ "$DISTRO_BINARY_COMPAT" == "void" ];then #20230914
- if [ ! -e /tmp/petget/void-db-updated-flg ];then
-  ping -4 -c 1 -w 5 -q google.com
-  if [ $? -ne 0 ];then
-   E1="$(gettext 'PKGget requires Internet access')"
-   popup "background=#ffa0a0 terminate=ok process=wait level=top|<big>${E1}</big>"
-   exit
+ #easyvoid built from woofQ does not use xbps; in that case must update db...
+ if [ ! -e /usr/bin/xbps-install ];then #20240228
+  if [ ! -e /tmp/petget/void-db-updated-flg ];then
+   ping -4 -c 1 -w 5 -q google.com
+   if [ $? -ne 0 ];then
+    E1="$(gettext 'PKGget requires Internet access')"
+    popup "background=#ffa0a0 terminate=ok process=wait level=top|<big>${E1}</big>"
+    exit
+   fi
+   W1="$(gettext 'As Void Linux is a rolling release, need to update the package database every time run PKGget. Please wait...')"
+   popup "background=#ffd8e6 level=top|<big>${W1}</big>"
+   /usr/local/petget/0setup q
+   sync
+   killall popup
+   touch /tmp/petget/void-db-updated-flg
   fi
-  W1="$(gettext 'As Void Linux is a rolling release, need to update the package database every time run PKGget. Please wait...')"
-  popup "background=#ffd8e6 level=top|<big>${W1}</big>"
-  /usr/local/petget/0setup q
-  sync
-  killall popup
-  touch /tmp/petget/void-db-updated-flg
  fi
 fi
